fix throughly

+++ vllm/model_executor/layers/fused_moe/modular_kernel.py
@@ -983,7 +983,7 @@ class FusedMoEModularKernel(torch.nn.Module):
             assert num_chunks == 0
             workspace13 = None
             workspace2 = None
-            fused_out = torch.empty_like(a1q)
+            fused_out = torch.empty_like(a1q, dtype=torch.bfloat16)
         else:
             assert num_chunks > 0
             workspace13, workspace2, fused_out = self._allocate_buffers(
